==> 2_students_files.ipynb <==
import pandas as pd
import numpy as np
import glob

## Astrtegia file
students = pd.read_excel("C:/Users/yoave/Documents/Data Projects/הכשרה מקצועית_יואב/Data/Students data/raw/students_10-21.xlsx")

## Print file info
uniques = students["id"].nunique()
rows = students.shape[0]
no_date = students.loc[(students["end_year"]==1900)]

print("Number of unique students: ", uniques)

print("Number of rows: ", rows)

print("Percentage of Students with one hahshra: " , round(100-uniques/rows*100, 2), "%")

nehiga = students.loc[(students["eshcol"] == 'נהיגה ')]
nehiga = nehiga.loc[(nehiga["end_year"] == 2018) | (nehiga["end_year"] == 2019)]
print(nehiga.shape)

## If end_date is null return end_date_p
students["end_date"] = np.where(students["end_date"].isnull(), students["end_date_p"], students["end_date"])

## replace 2206,2109,2104,2121 with 0
students["end_year"] = students["end_year"].replace([2206,2109,2104,2121],0)
students = students.loc[(students["end_year"] != 0)]

## Keep only last end_date
date_cols=["end_date"]
students[date_cols] = students[date_cols].apply(pd.to_datetime, format='%d/%m/%Y')

## Keep relevant maslulim
### 5=shikum, 6=shabas , 13 = hashmal_mesuyag, 15 = OJT, [21,31,67] = hishtalmut, 34 = patuah , 33= yarok , 37= hashlamat haskala
drop_maslulim = [5,6,13,15,21,31,67,34,33,37]
droped = students[students['maslul'].isin(drop_maslulim)]
## droped shape
droped["id"].nunique()
students = students[~students['maslul'].isin(drop_maslulim)]

## Keep relevant status
drop_statusim = ["עזב","בוגר - מסיים","עבר לקורס חדש"]
droped_status = students[students['student_status'].isin(drop_statusim)]
students = students[~students['student_status'].isin(drop_statusim)]

## Keep only first course data for each id
df = students.sort_values(by=["id",'end_date']).drop_duplicates(subset='id', keep='first')

nehiga = df.loc[(df["eshcol"] == 'נהיגה ')]
nehiga = nehiga.loc[(nehiga["end_year"] == 2018) | (nehiga["end_year"] == 2019)]
print(nehiga.shape)

df.to_csv("Data/Students data/students_processed_2.csv", index=False)
