==> 3_merge_stud_batla.ipynb <==
import pandas as pd
import numpy as np
import datetime


students = pd.read_csv("Data/Students data/students_processed_2.csv")
batla = pd.read_csv("Output/BONUS_HASAM_202206233_35.csv")

students["end_date"] = students["end_date"].apply(pd.to_datetime, format='%Y-%m-%d')

## keep columns
keep_cols = ["id","gender","end_date","end_year","mosad_id","eshcol","group"]
students = students[keep_cols]

## drop Unnamed: 0
batla = batla.drop(columns=["Unnamed: 0"])

## Merge Leftovers
merged = pd.merge(students, batla, how="outer", on="id", indicator=True)
left_only = merged[merged["_merge"] == "left_only"]

left_only["end_year"].value_counts(sort=True)

df = pd.merge(batla, students, on="id", how="left")

nehiga = students.loc[(students["eshcol"] == 'נהיגה ')]
nehiga = nehiga.loc[(nehiga["end_year"] == 2018) | (nehiga["end_year"] == 2019)]
nehiga["id"].nunique()

## Filter after 2016
df = df.loc[(df["end_year"] >= 2016)]

df["date"] =  pd.to_datetime(df["date"])
#df["end_date"] =  pd.to_datetime(df["end_date"])

def inter(x):

    if x["date"].year == x["end_date"].year and x["date"].month == x["end_date"].month:
        return 1
    else:
        return 0

df["intervention"] = df.apply(inter, axis=1)

# Index monthes before and after intervention
df = df.sort_values(by=['id','date'], ascending=False)
df['intervention_index_pre'] = df.groupby('id',group_keys=False)['intervention'].apply(lambda x: (x - x.cumsum()).where(x == 1, lambda x: x.cumsum()))

df = df.sort_values(by=['id','date'], ascending=True)
df['intervention_index_post'] = df.groupby('id',group_keys=False)['intervention'].apply(lambda x: (x - x.cumsum()).where(x == 1, lambda x: x.cumsum()))
df["intervention_index_post"] = df["intervention_index_post"]*-1

def intervention(x):
    if x["intervention_index_pre"] < 0:
        return x["intervention_index_pre"]
    if x["intervention_index_pre"] == 0:
        return x["intervention_index_post"]
    else:
        return 0

df["intervention_index"] = df.apply(intervention, axis=1)
df= df.drop(columns=["intervention_index_pre","intervention_index_post"])

## Drop observation finished after batla data range
df = df.loc[(df["intervention"]!=0) | (df["intervention_index"]!=0)]

df.to_csv("Output/merge_stud_batla.csv", index=False)
